Adds sbt-typelevel

+++ aws/src/main/scala/dev/posco/hiona/aws/LambdaDeploy.scala
@@ -1,12 +1,24 @@
+/*
+ * Copyright 2022 devposco
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package dev.posco.hiona.aws
 
 import cats.data.{Validated, ValidatedNel}
 import cats.effect.{Blocker, ContextShift, ExitCode, IO, IOApp, Resource, Timer}
-import dev.posco.hiona.{IOEnv, LazyToString}
-import java.nio.file.Path
-import com.monovore.decline.{Argument, Command, Opts}
 import com.amazonaws.PredefinedClientConfigurations
-import com.amazonaws.services.lambda.{AWSLambda, AWSLambdaClientBuilder}
 import com.amazonaws.services.lambda.model.{
   CreateFunctionRequest,
   DeleteFunctionRequest,
@@ -17,11 +29,15 @@ import com.amazonaws.services.lambda.model.{
   Runtime,
   VpcConfig
 }
+import com.amazonaws.services.lambda.{AWSLambda, AWSLambdaClientBuilder}
 import com.amazonaws.services.s3
+import com.monovore.decline.{Argument, Command, Opts}
+import dev.posco.hiona.{IOEnv, LazyToString}
 import io.circe.Json
-import io.circe.parser.decode
 import io.circe.jawn.CirceSupportParser
+import io.circe.parser.decode
 import java.nio.ByteBuffer
+import java.nio.file.Path
 import org.apache.commons.codec.binary.Base64
 import org.slf4j.LoggerFactory
 import scala.concurrent.duration.FiniteDuration
@@ -649,7 +665,7 @@ object LambdaDeploy {
       val cconf =
         Option(bldr.getClientConfiguration)
           .getOrElse(PredefinedClientConfigurations.defaultConfig)
-          .withConnectionMaxIdleMillis(timeout)
+          .withConnectionMaxIdleMillis(timeout.toLong)
           .withConnectionTimeout(timeout)
           .withSocketTimeout(timeout)
           .withRequestTimeout(timeout)
