+++ core/src/test/scala/dev/posco/hiona/DurationLaws.scala
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2022 devposco
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package dev.posco.hiona
 
 import org.scalacheck.Gen
@@ -56,7 +72,7 @@ class DurationLaws extends munit.ScalaCheckSuite {
 
   property("(d * n)/n == d, or d * n is infinite") {
     forAll(genDuration, Gen.choose(1, Int.MaxValue)) { (d, n1) =>
-      val n = n1 max 1
+      val n = (n1 max 1).toLong
       val d1 = d * n
       val d2 = d1 / n
       if (d1.isInfinite) assert(d2.isInfinite)
